Make BufferedWriter an internal class again, now that it has been simplified
