+++ Firestore/core/src/firebase/firestore/remote/grpc_stream.cc
@@ -16,8 +16,8 @@
 
 #include "Firestore/core/src/firebase/firestore/remote/grpc_stream.h"
 
-#include <chrono>
-#include <future>
+#include <chrono>  // NOLINT(build/c++11)
+#include <future>  // NOLINT(build/c++11)
 
 #include "Firestore/core/src/firebase/firestore/remote/datastore.h"
 
@@ -63,6 +63,34 @@ using util::AsyncQueue;
 //   responsibility of the callers to invoke its methods in appropriate
 //   execution contexts.
 
+namespace internal {
+
+StreamWrite* BufferedWriter::EnqueueWrite(grpc::ByteBuffer&& write) {
+  queue_.push(write);
+  return TryStartWrite();
+}
+
+StreamWrite* BufferedWriter::TryStartWrite() {
+  if (queue_.empty() || has_active_write_) {
+    return nullptr;
+  }
+
+  has_active_write_ = true;
+  grpc::ByteBuffer message = std::move(queue_.front());
+  queue_.pop();
+  return StreamOperation::ExecuteOperation<StreamWrite>(
+      stream_, call_, firestore_queue_, std::move(message));
+}
+
+StreamWrite* BufferedWriter::DequeueNextWrite() {
+  has_active_write_ = false;
+  return TryStartWrite();
+}
+
+}  // namespace internal
+
+using internal::BufferedWriter;
+
 GrpcStream::GrpcStream(
     std::unique_ptr<grpc::ClientContext> context,
     std::unique_ptr<grpc::GenericClientAsyncReaderWriter> call,
@@ -79,7 +107,7 @@ GrpcStream::GrpcStream(
 
 GrpcStream::~GrpcStream() {
   HARD_ASSERT(operations_.empty(),
-              "GrpcStream is being destroyed without a call to Finish");
+              "GrpcStream is being destroyed without proper shutdown");
 }
 
 void GrpcStream::Start() {
@@ -106,7 +134,8 @@ void GrpcStream::Write(grpc::ByteBuffer&& message) {
 
 void GrpcStream::Finish() {
   if (state_ == State::NotStarted) {
-    HARD_ASSERT(operations_.empty(), "Non-started stream has pending operations");
+    HARD_ASSERT(operations_.empty(),
+                "Non-started stream has pending operations");
     state_ = State::Finished;
     return;
   }
@@ -118,11 +147,14 @@ void GrpcStream::Finish() {
   // cancellation has to be called, or else the read would hang forever, and
   // finish operation will never get completed.
   //
-  // On the other hand, when an operation fails, cancellation should not be
+  // (on the other hand, when an operation fails, cancellation should not be
   // called, otherwise the real failure cause will be overwritten by status
-  // "canceled".
+  // "canceled".)
   context_->TryCancel();
-  Execute<ClientInitiatedFinish>();  // TODO: is it necessary?
+  // TODO(varconst): is issuing a finish operation necessary in this case? We
+  // don't care about the status, but perhaps it will make the server notice
+  // client disconnecting sooner?
+  Execute<ClientInitiatedFinish>();
 
   FastFinishOperationsBlocking();
 }
@@ -131,11 +163,19 @@ void GrpcStream::FastFinishOperationsBlocking() {
   // TODO(varconst): reset buffered_writer_? Should not be necessary, because it
   // should never be called again after state_ == State::Finished.
 
+  HARD_ASSERT(state_ == State::Finishing,
+              "Fast-finishing operations must only be done when the stream is "
+              "finishing");
+
   for (auto operation : operations_) {
+    // `GrpcStream` cannot cancel the completion of any operations that might
+    // already have been enqueued on the Firestore queue, so instead turn those
+    // completions into no-ops.
     operation->UnsetObserver();
   }
 
   for (auto operation : operations_) {
+    // This is blocking.
     operation->WaitUntilOffQueue();
   }
   operations_.clear();
@@ -150,10 +190,17 @@ bool GrpcStream::WriteAndFinish(grpc::ByteBuffer&& message) {
 
   bool did_last_write = false;
   StreamWrite* last_write_operation = BufferedWrite(std::move(message));
+  // Only bother with the last write if there is no active write at the moment.
   if (last_write_operation) {
     last_write_operation->UnsetObserver();
+    // Empirically, the write normally takes less than a millisecond to finish
+    // (both with and without network connection), and never more than several
+    // dozen milliseconds. Nevertheless, ensure `WriteAndFinish` doesn't hang if
+    // there happen to be circumstances under which the write may block
+    // indefinitely (in that case, rely on the fact that canceling GRPC call
+    // makes all pending operations come back from the queue quickly).
     auto status =
-        last_write_operation->WaitUntilOffQueue(std::chrono::milliseconds(100));
+        last_write_operation->WaitUntilOffQueue(std::chrono::milliseconds(500));
     if (status == std::future_status::ready) {
       RemoveOperation(last_write_operation);
       did_last_write = true;
@@ -195,8 +242,7 @@ GrpcStream::MetadataT GrpcStream::GetResponseHeaders() const {
 
 void GrpcStream::OnStart() {
   HARD_ASSERT(state_ == State::Starting,
-              "Expected to be in 'Starting' state "
-              "when OnStart is invoked");
+              "Expected to be in 'Starting' state when OnStart is invoked");
   state_ = State::Open;
 
   if (SameGeneration()) {
@@ -236,8 +282,9 @@ void GrpcStream::OnOperationFailed() {
     return;
   }
 
+  state_ = State::Finishing;
+
   if (SameGeneration()) {
-    state_ = State::Finishing;
     Execute<RemoteInitiatedFinish>();
   } else {
     // The only reason to finish would be to get the status; if the observer is
@@ -262,8 +309,6 @@ void GrpcStream::OnFinishedByClient() {
 void GrpcStream::RemoveOperation(const StreamOperation* to_remove) {
   auto found = std::find(operations_.begin(), operations_.end(), to_remove);
   HARD_ASSERT(found != operations_.end(), "Missing StreamOperation");
-  // Note that the operation might have contained the last reference to this
-  // stream, so this call might trigger `GrpcStream::~GrpcStream`.
   operations_.erase(found);
 }
